+++ assets/js/warehouse.js
@@ -520,7 +520,11 @@ var openMenu = function() {
 function submenuOffset() {
 	$(".adminmenu .menu-top").mouseover(function() {
 		var submenu = $(this).next(".wp-submenu"),
-			moveup = $("#footer").offset().top - $(this).offset().top - submenu.outerHeight();
+			offsetTop = $("#footer").offset().top;
+
+		offsetTop += ($("#footer").css('bottom') != '0px' ? window.innerHeight : 0);
+		moveup = offsetTop - $(this).offset().top - submenu.outerHeight();
+		console.log($("#footer").css('bottom'), offsetTop, $(this).offset().top, submenu.outerHeight(), moveup);
 		submenu.css("margin-top", (moveup < 0 ? moveup : 0) + 'px');
 	});
 }